changed datatype to int; fixed time measurement gpu;

+++ OpenCLMatrix/Kernel.cl
@@ -1,5 +1,5 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
-__kernel void OpenCLMatrix(__global double *A, __global double *b, __global double *c, const unsigned int n)
+__kernel void OpenCLMatrix(__global int *A, __global int *b, __global int *c, const unsigned int n)
 {
     //Get our global thread ID
     int id = get_global_id(0);
